src/main/java/com/springbootaws/book/web/HelloController.java
package com.springbootaws.book.web;

import com.springbootaws.book.web.dto.HelloResponseDto;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

@RestController //컨트롤러를 json으로 반환하는 컨트롤러
public class HelloController {
    @GetMapping("/hello")
    public String hello() {
        return "helloaws";
    }

    @GetMapping("/hello/dto")
    public HelloResponseDto helloDto(@RequestParam("name") String name,
                                     @RequestParam("amount") int amount) {
        return new HelloResponseDto(name, amount);
    }

}
